Update Transactions Controller, Model, remove Refunds Model, View, Controller
We store the refund_transactionid (the transactionid of the order transaction) so that we can retrieve the original order information, as well as the refund amount of the refund transaction, so there is no real need to store duplicate info in a refunds table.

Update to View/Transactions/view.ctp for the Order info for the Refund.

Signed-off-by: Stephen Mareches <stephen.mareches@payscape.com>